Remove dead web package helpers
