+++ apps/web/package.json
@@ -13,7 +13,6 @@
   },
   "dependencies": {
     "@sentry/react": "^10.62.0",
-    "fflate": "^0.8.3",
     "heic-to": "^1.5.2",
     "lottie-web": "^5.13.0",
     "qrcode.react": "^4.2.0",
