+++ apps/web/src/screens/settings/workspace/WorkspaceExportScreen.tsx
@@ -11,7 +11,6 @@ import type {
   WorkspacePackageExportPreviewResponse,
   WorkspacePackageExportRequest,
 } from "../../../types";
-import { triggerBlobDownload } from "../../../workspaceExport";
 import { SettingsShell } from "../SettingsShared";
 
 type PackageMetadataRow = Readonly<{
@@ -25,6 +24,36 @@ type PackageExportPreviewIdentity = Readonly<{
 }>;
 
 type NumberFormatter = (value: number, options?: Readonly<Intl.NumberFormatOptions>) => string;
+
+type WorkspaceExportUrlApi = Readonly<{
+  createObjectURL: (object: Blob) => string;
+  revokeObjectURL: (url: string) => void;
+}>;
+
+type TriggerBlobDownloadParams = Readonly<{
+  blob: Blob;
+  filename: string;
+  document: Document;
+  urlApi: WorkspaceExportUrlApi;
+}>;
+
+function triggerBlobDownload(params: TriggerBlobDownloadParams): void {
+  const { blob, filename, document, urlApi } = params;
+  if (document.body === null) {
+    throw new Error(`Document body is unavailable for workspace download: filename=${filename}`);
+  }
+
+  const objectUrl = urlApi.createObjectURL(blob);
+  const anchor = document.createElement("a");
+  anchor.href = objectUrl;
+  anchor.download = filename;
+  anchor.rel = "noopener";
+  anchor.style.display = "none";
+  document.body.append(anchor);
+  anchor.click();
+  anchor.remove();
+  urlApi.revokeObjectURL(objectUrl);
+}
 
 function buildDefaultWorkspacePackageExportPreviewRequest(): WorkspacePackageExportRequest {
   return {
